Separate io-text and nvtext pytests into different files
