--- python/cudf/cudf/tests/text/test_subword_tokenizer.py
@@ -1,1 +1,0 @@
-# Copyright (c) 2023, NVIDIA CORPORATION.
